refactor: split util.ts into utils directory with modules that group utilities
