+++ src/utils/async.ts
@@ -1,0 +1,5 @@
+export function sleep(time: number) {
+  return new Promise((resolve) => {
+    setTimeout(resolve, time);
+  });
+}
